plugins_dev/mc-notifier/pyproject.toml
[project]
name = "airflow-provider-minecraft"
description = "An Apache Airflow provider to send notification to a Minecraft server"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [ "apache-airflow" ]
dynamic = [ "version" ]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.entry-points.apache_airflow_provider]
provider_info = "minecraft_provider.__init__:get_provider_info"

[tool.setuptools]
include-package-data = false

[tool.setuptools.dynamic]
version = { attr = "minecraft_provider.__version__" }